+++ R/jsBasicGadget.R
@@ -130,6 +130,7 @@ jsBasicGadget <- function(data, nfactor.limit = 20) {
                                            kaplanUI("kaplan")
                                          ),
                                          mainPanel(
+                                           optionUI("kaplan"),
                                            withLoader(plotOutput("kaplan_plot"), type="html", loader="loader6"),
                                            ggplotdownUI("kaplan")
                                          )