+++ R/jsSurveyGadget.R
@@ -137,6 +137,7 @@ jsSurveyGadget <- function(data, nfactor.limit = 20) {
                                            kaplanUI("kaplan")
                                          ),
                                          mainPanel(
+                                           optionUI("kaplan"),
                                            withLoader(plotOutput("kaplan_plot"), type="html", loader="loader6"),
                                            ggplotdownUI("kaplan")
                                          )